+++ apps/client/src/shared/api/getPosts.ts
