+++ apps/client/src/shared/ui/index.ts
@@ -1,4 +1,4 @@
 export { default as Dropdown } from "./dropdown";
 export { default as File } from "./file";
 export { default as Textarea } from "./textarea";
-export { default as Post } from "./post/post";
+export { default as Post } from "./post";
