+++ apps/client/src/shared/ui/post/index.tsx
